[server] | refactor(Videos pipe): getVideosPipe 수정
