+++ packages/server/src/common/pipes/get-videos.pipe/dto.ts
@@ -1,0 +1,6 @@
+export class GetVideosPipeDto {
+  public constructor(
+    public readonly page: string,
+    public readonly sort: string,
+  ) {}
+}
